streaming-microservice-main/movie_list/requirements.txt
grpcio-tools ~= 1.30